chore: define code of conduct
A code of conduct defines standards for how to engage in a community. It
signals an inclusive environment that respects all contributions. It
also outlines procedures for addressing problems between members of your
project's community.

See: https://opensource.guide/code-of-conduct/
See: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project
See: https://www.contributor-covenant.org/