chore(commits): enforce body as sentences and allow long footers
This allows URL references in the footer tagged with `See: ` and typical
git trailers. The _body as sentences_ part is to enforce the optional
body to read like paragraphs and sentences. I find this makes for an
easier reading, since this can be desirably verbose at times.

See: https://chris.beams.io/posts/git-commit/