feat: create linter binary named `ingenots`
This is inspired by `toddbluhm/ts-standard` and based on
`standard/standard-engine`. The whole point of this repository is to
create a set of JS/TypeScript linting/formatting guidelines bundled as a
binary that is easy to include and integrate in any project with minimal
configuration.

See: https://github.com/standard/standard-engine
See: https://github.com/toddbluhm/ts-standard